requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
annotated-types==0.5.0
antiorm==1.2.1
anyio==3.7.1
async-timeout==4.0.3
attrs==23.1.0
blinker==1.6.2
certifi==2023.7.22
charset-normalizer==3.3.0
click==8.1.7
colorama==0.4.6
dataclasses-json==0.6.1
DateTime==5.2
dnspython==2.4.2
Flask==2.3.3
Flask-Cors==4.0.0
frozenlist==1.4.0
greenlet==3.0.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
jsonpatch==1.33
jsonpointer==2.4
langsmith==0.0.41
MarkupSafe==2.1.3
marshmallow==3.20.1
multidict==6.0.4
mypy-extensions==1.0.0
numexpr==2.8.7
numpy==1.26.0
openai==0.28.1
packaging==23.2
pydantic==2.4.2
pydantic_core==2.10.1
PyJWT==2.8.0
pymongo==4.5.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
PyYAML==6.0.1
requests==2.31.0
six==1.16.0
sniffio==1.3.0
soupsieve==2.5
SQLAlchemy==2.0.21
tenacity==8.2.3
tqdm==4.66.1
typing-inspect==0.9.0
typing_extensions==4.8.0
urllib3==2.0.6
Werkzeug==2.3.7
yarl==1.9.2
zope.interface==6.0
